rename wp_parser.tab.c to wp_parser.tab.cpp
